# Speed up XORMappedAddress by inlining xorBytes