… xor.go
… xor_test.go