#### Description

Since xorBytes is only ever called for very small slices, the expensive part of it is the function call. Benchmark results on
linux/amd64:

```
name                        old time/op    new time/op    delta
XORMappedAddress_AddTo-8      43.5ns ±12%    33.1ns ± 1%  -23.91%  (p=0.000 n=10+9)
XORMappedAddress_GetFrom-8    26.2ns ± 7%    19.6ns ±20%  -25.36%  (p=0.000 n=10+9)
```

#### Reference issue
Closes #116.