@jech 2022-07-11
See also #116, which does the exact opposite ;-)