@greatroar 2022-07-30
I saw that, but then the question is whether you want all users to pull in a dependency with five asm implementations that are actually slower than a simple loop.