@stv0g 2022-11-12
This looks like a prime example of premature optimization to me.

I support @jech proposal to centralize the implementation in the transport package.