@jech 2022-11-12
I really don't care either way ­— either this PR or #116 are perfectly fine with me.  On the other hand, I'd like to see one or the other applied, since I'm uncomfortable with the code duplication that's happening in the current code.