@stv0g 2022-11-12
I've reviewed and merged #116. Hence this PR is obsolete.